+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## v2.0.21
 - Popup notification extension has been created
 - Fix next video timer's -1 value issue
+- Decrease midroll interval to 10000ms
 
 ## v2.0.20.4
 - Avoid duplicate ad request when play the video